+++ lib/rubocop/cop/style/struct_inheritance.rb
@@ -20,6 +20,8 @@ module RuboCop
       #     end
       #   end
       class StructInheritance < Cop
+        include RangeHelp
+
         MSG = "Don't extend an instance initialized by `Struct.new`. " \
               'Use a block to customize the struct.'
 
@@ -33,6 +35,48 @@ module RuboCop
           {(send (const nil? :Struct) :new ...)
            (block (send (const nil? :Struct) :new ...) ...)}
         PATTERN
+
+        def autocorrect(class_node)
+          lambda do |corrector|
+            return nil if struct_definition_has_block?(class_node)
+
+            remove_class_keyword(class_node, corrector)
+
+            corrector.replace(class_node.loc.operator, '=')
+
+            transform_class_body(class_node, corrector)
+          end
+        end
+
+        private
+
+        def struct_definition_has_block?(class_node)
+          class_node.parent_class.is_a?(RuboCop::AST::BlockNode)
+        end
+
+        def remove_class_keyword(class_node, corrector)
+          corrector.remove(
+            range_with_surrounding_space(
+              range: class_node.loc.keyword,
+              side: :right
+            )
+          )
+        end
+
+        def transform_class_body(class_node, corrector)
+          body = class_node.body
+
+          if body
+            corrector.insert_after(class_node.parent_class.source_range, ' do')
+          else
+            corrector.remove(
+              range_with_surrounding_space(
+                range: class_node.loc.end,
+                side: :right
+              )
+            )
+          end
+        end
       end
     end
   end